renamed 2 class names, moved all history controllers under AdminController package
