+++ src/frontend/BasicFXML/CategoryFilter.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.VBox?>
 
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.161" xmlns:fx="http://javafx.com/fxml/1" fx:controller="frontend.HistoryControllers.CategoryFilterController">
+<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.161" xmlns:fx="http://javafx.com/fxml/1" fx:controller="frontend.AdminControllers.HistoryCategoryFilterController">
     <ScrollPane layoutX="100.0" layoutY="87.0" prefHeight="200.0" prefWidth="401.0">
         <VBox fx:id="scrollCategoryFilter" minHeight="0.0" minWidth="0.0" prefHeight="200.0" prefWidth="200.0" />
     </ScrollPane>
