+++ src/frontend/BasicFXML/HistoryDelete.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 
 <AnchorPane prefHeight="100.0" prefWidth="400.0" xmlns="http://javafx.com/javafx/8.0.161"
-            xmlns:fx="http://javafx.com/fxml/1" fx:controller="frontend.HistoryControllers.HistoryDeleteController">
+            xmlns:fx="http://javafx.com/fxml/1" fx:controller="frontend.AdminControllers.HistoryDeleteController">
    <Button layoutX="58.0" layoutY="44.0" mnemonicParsing="false" onAction="#onCancelAction" text="Cancel"/>
    <Button layoutX="283.0" layoutY="44.0" mnemonicParsing="false" onAction="#onConfirmAction" text="Confirm"/>
    <Label layoutX="164.0" layoutY="14.0" text="Are you sure?"/>
