+++ src/frontend/BasicFXML/HistoryDetails.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.control.TableView?>
 <?import javafx.scene.layout.AnchorPane?>
 
-<AnchorPane prefHeight="720.0" prefWidth="1800.0" xmlns="http://javafx.com/javafx/8.0.161" xmlns:fx="http://javafx.com/fxml/1" fx:controller="frontend.HistoryControllers.HistoryDetailsController">
+<AnchorPane prefHeight="720.0" prefWidth="1800.0" xmlns="http://javafx.com/javafx/8.0.161" xmlns:fx="http://javafx.com/fxml/1" fx:controller="frontend.AdminControllers.HistoryDetailsController">
     <TableView fx:id="table_parts" editable="true" layoutX="14.0" layoutY="104.0" prefHeight="469.0" prefWidth="1783.0">
       <columns>
           <TableColumn fx:id="col_customer_name" editable="false" prefWidth="69.99996948242188" sortable="false" text="Customer" />
